Fix propagate exit code (-p)
Result now shows correct exit code even when -p was not supplied. -p only
activates runlim itself to also return the exit code of the program. Previously,
the ok case would overwrite the status of the program with the injected status
of runlim, even if this would not have been required.